+++ wallet/CHANGELOG.md
@@ -25,6 +25,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Resync outputs if a transaction got confirmed between syncing outputs and pending transactions to prevent not having unspent outputs afterwards;
 
+### Fixed
+
+- `BaseCoinBalance::available` can't underflow anymore;
+
 ## 1.0.0-rc.6 - 2023-03-09
 
 ### Added