+++ wallet/CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Resync outputs if a transaction got confirmed between syncing outputs and pending transactions to prevent not having unspent outputs afterwards;
+- Make `Storage::get` generic over a `T: Deserialize` and return a `T`, avoiding always having to deserialize after;
 
 ## 1.0.0-rc.6 - 2023-03-09